requirements.txt
pip==20.2.3
Django==3.1
asgiref==3.2.10
pytz==2020.1
sqlparse==0.3.1
psycopg2-binary==2.8.5
django-environ==0.4.5
requests==2.24.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
urllib3==1.25.10
autopep8==1.5.4
pycodestyle==2.6.0
toml==0.10.1
pylint==2.5.3
astroid==2.4.2
isort==4.3.21
mccabe==0.6.1
lazy-object-proxy==1.4.3
wrapt==1.12.1
flake8==3.8.3
pyflakes==2.2.0
six==1.15.0
djangorestframework==3.11.1